contract::core::aliases
Type Alias ProvResponse

pub type ProvResponse = Response;

Aliased Type

struct ProvResponse {
    pub messages: Vec<SubMsg>,
    pub attributes: Vec<Attribute>,
    pub events: Vec<Event>,
    pub data: Option<Binary>,
}

Fields
messages: Vec<SubMsg>

Optional list of messages to pass. These will be executed in order. If the ReplyOn variant matches the result (Always, Success on Ok, Error on Err), the runtime will invoke this contract’s reply entry point after execution. Otherwise, they act like “fire and forget”. Use SubMsg::new to create messages with the older “fire and forget” semantics.

attributes: Vec<Attribute>

The attributes that will be emitted as part of a “wasm” event.

More info about events (and their attributes) can be found in Cosmos SDK docs.

events: Vec<Event>

Extra, custom events separate from the main wasm one. These will have wasm- prepended to the type.

More info about events can be found in Cosmos SDK docs.

data: Option<Binary>

The binary payload to include in the response.