Module contract::mint

Functions

add_nft_holder

Gets a holder_id by adding a new holder

generate_metata

Generates metadata from given metadadata string, price and comission

get_mint_dicts

Gets the dicts needed to be worked with in mint entrypoint

get_mint_runtime_args

Returns the runtime args needed for mint entrypoint to run

get_new_token_id

Gets a new token_id from the contract state and returns is as a u64

mint

Mint Entrypoint of the contract