Function contract::mint::add_nft_holder

fn add_nft_holder(
    holders_cnt_uref: URef, 
    owners_dict_uref: URef, 
    holder_by_id_uref: URef, 
    reciver: String, 
    _token_id: u64, 
    amount: u64
) -> u64

Gets a holder_id by adding a new holder

Adds the minted NFT to the holders list of the owner account, It would search for the holder_id that corresponds with the token_id, and if it found it, It would modify the amount of it and add the amount to it. If it failed to find the holder_id , it would use holders_cnt+1 as new holder_id, and modify holders_cnt to holders_cnt+1, and finally it would return the final_holder_id (the new created one or existing one based on the situation).