Module contract::ndpc_types

Structs

ApprovedNFT

This struct is used to store the approved NFTs (approved to publish)

MetadataHash

Holds the hash of the metadata

NFTHolder

NFTHolder : an amount and a token_id which identifies an NFT

NftMetadata

Metadata of the NFT, including name, uri, checksum, price and comission

PublishRequest

This struct is used to store publish requests

U64list

a simple wrapper for a set of u64

Constants

METADATA_HASH_LENGTH

Hash-len of the metadata, its blake2b so it would be 32 bytes

Traits

AsStrized

Converts the given object to String, Its used to convert the AccountHash to base16 encoded string

FromStringize

Converts the given Strign to the type, used to convert hex encoded string to accounthash