Module contract::ndpc_utils

Structs

PaymentDetails

Functions

calculate_payment

Calculates the amount of motes that need to be transfered to producer, and droplinked account

contract_package_hash

Returns the contractPackage hash

get_approved_holder_by_id

Gets a approved_id and returns the ApprovedNFT related to that, gets the ApprovedNFT from the approved dict

get_droplinked_account

Simply returns the ratioVerifier of the contract

get_fee

A shortcut function, which returns the fee that was set in the contract state during installation process

get_holder_by_id

Gets holder_id and returns the NFTHolder object that has the holder_id in the holders_dict

get_holder_ids

A shortcut function which returns the holder_ids which a account owns, as a Option

get_holders_cnt

A shortcut function which returns the holders_cnt as a u64

get_named_key_by_name

A shortcut function to get a dictionary by its name, and return its URef value

get_nft_metadata

Gets token_id and returns the NFTMetadata object related to it

get_ratio_verifier

A getter function, which gets the ratio verifier account from the contract state, and return it as a PublicKey

get_request_by_id

Gets request_id, and returns the PublisRequest related to it, gets the publishRequest from the request_objects dict

verify_signature

Verify a signature of a message, which is signed by the given publicKey