contract::storage
Module loan_pool_collateral

Constants

COLLATERAL

Functions

exists

function defining if the marker has been added to the pool.

get

Get Collateral from storage.

get_all_states

get_with_state

remove

This function is declared as a public function named remove. This means it can be called from other modules. It expects two parameters: storage: a mutable reference to a trait object of dyn Storage. Traits are a way to group method signatures that define a specific behavior in a generic way. dyn keyword is used for dynamic dispatch. collateral: an immutable reference to an instance of LoanPoolMarkerCollateral.

set

Save Collateral to storage.