contract::storage::loan_pool_collateral
Function exists

pub fn exists(storage: &dyn Storage, marker_address: Addr) -> bool

function defining if the marker has been added to the pool.

Parameters

storage: It takes a shared reference to a dynamic Storage trait object.

marker_address: An instance of Addr which is likely used to represent a specific address or location within the storage.

This function checks if a certain marker address exists within the storage by calling the has method on the COLLATERAL variable.

Returns

A boolean value. Returning true if the marker_address exists in the COLLATERAL, otherwise false.

Note

It’s important to remember the behavior of the function could differ considerably based on the specific implementation of the COLLATERAL, Storage trait and has method.