contract::storage::whitelist_contributors_store
Function get_whitelist_contributors

pub fn get_whitelist_contributors(storage: &dyn Storage) -> Vec<Addr>

This function retrieves the list of contributors from the whitelist in the given storage.

Arguments

storage - A storage object implementing the Storage trait. This is where the function looks for the whitelist of contributors.

Returns

A Vec<Addr> representing the list of contributors. If the whitelist cannot be loaded from storage, this function will return an empty Vec<Addr>.

Examples

Panics

This function will panic if an error occurs while loading the whitelist from storage and will return an empty vector as default.

Safety

This function assumes the storage provided can correctly and safely load the items requested.