contract::storage::whitelist_contributors_store
Function save_contributors

pub fn save_contributors(
    storage: &mut dyn Storage,
    new_contributors: Vec<Addr>,
) -> Result<(), ContractError>

Adds a list of new contributors to the existing whitelist.

Arguments

storage - A mutable reference to the contract’s storage.

new_contributors - A vector of addresses (contributors) to be added to the whitelist.

Returns

A Result which is:

Ok(()) on success.

Err(ContractError) on failure, where ContractError is an enum defined within the contract to handle possible error cases.

Errors

Will return Err(ContractError::Std(StdError)) where StdError is the error returned from the storage API if it fails to load or save data from/to storage.

Example

let new_contributors = vec![Addr::unchecked("addr1"), Addr::unchecked("addr2")];
save_contributors(deps.storage, new_contributors);