contract::util::settlement
Function timestamp_is_expired

pub fn timestamp_is_expired(
    storage: &dyn Storage,
    time: &Timestamp,
) -> Result<bool, ContractError>

Checks if the timestamp is expired.

Parameters

storage A reference to the storage object in the dependencies provided by the cosmwasm framework.

time A reference to the timestamp that will be checked for expiration.

Examples